% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_surv_times.R
\name{gen_surv_times}
\alias{gen_surv_times}
\title{Generate survival times}
\usage{
gen_surv_times(invchaz, mu = log(1), data, coxphmod = NULL)
}
\arguments{
\item{invchaz}{the inverse cumulative (baseline) hazard rate to be used
for generating survival times. Must take vector inputs!}

\item{mu}{the true hazard ratio used to generate survival times.}

\item{data}{an integer number of survival times to generate or
(in combination with coxphmod): a \code{data.frame} containing
subject covariates in named columns.}

\item{coxphmod}{(optional) a cox proportional hazards regression model as produced by
the function \code{\link[survival:coxph]{coxph()}}. Standard practice: \cr
\code{coxph(Surv(survtime, censorid) ~ covariates, data = data)}. \cr
Alternatively, a list with:
\itemize{
\item $formula (~ covariates)
\item $coefficients (named vector specifying risk adjustment coefficients
for covariates - names must be the same as in $formula and colnames of \code{data}).
}}
}
\value{
A vector of survival times from subject entry time.
}
\description{
Generate survival times according to hazard rate
\eqn{h(t) \exp(\mu)} with \eqn{h(t)} the hazard rate associated with the
specified inverse cumulative hazard rate \code{invchaz} and \eqn{\mu} the
specified true hazard ratio \code{mu}. See Bender et al. (2005).
}
\details{
Sometimes it is desirable to generate survival times from an
increased hazard rate \deqn{h(t, \mu) = h_0(t) e^\mu}{h(t, \mu) = h_0(t) exp(\mu)}
with \eqn{h_0}{h_0} the baseline hazard rate. We call \eqn{e^\mu}{exp(\mu)} the true hazard ratio.
}
\examples{
gen_surv_times(invchaz = function(t) inv_chaz_exp(t, lambda = 0.01), data = 5)
}
\references{
Bender, R., Augustin, T., & Blettner, M. (2005).
Generating survival times to simulate Cox proportional hazards models.
Statistics in medicine, 24(11), 1713-1723. \doi{10.1002/sim.2059}
}
\seealso{
Other utils: 
\code{\link{calc_risk}()},
\code{\link{exp_hazards}},
\code{\link{gen_arriv_times}()},
\code{\link{runlength}()}
}
\author{
Daniel Gomon
}
\concept{utils}
