% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr_summary.R
\name{summary.funnelplot}
\alias{summary.funnelplot}
\title{Summarizes (or elaborates on) S3 objects in this package.}
\usage{
\method{summary}{funnelplot}(object, ...)
}
\arguments{
\item{object}{S3 object to summarize}

\item{...}{extra parameters}
}
\value{
A list with:
\itemize{
\item \code{$call} The call used to obtain the input object,
\item \code{$'0.xx'} The detected instances at specified confidence level.
}
}
\description{
Prints the (name of the) instances performing worse than expected
in a "funnelplot" object at the specified confidence levels.
}
\section{Methods (by class)}{
\itemize{
\item \code{funnelplot}: Summarize instances detected by the funnelplot object
}}

\seealso{
, \code{\link[cgrcusum:funnelplot]{funnelplot}}
}
